MF-1343 CM Field is not resetting based on conditions
Changes to the fee indicator dynamic set logic to have a higher priority than others who call that action.  this will cause the change to be done before the change report is done.  also made a gv change as the form it was sending did not contain the admin fields